restructure done withowt nfa and nfa transformation
